WithTooltip moved to separate directory
