+++ packages/core/src/renderer/components/config-limit-ranges/limit-ranges.tsx
@@ -9,12 +9,12 @@ import "./limit-ranges.scss";
 import { withInjectables } from "@ogre-tools/injectable-react";
 import { observer } from "mobx-react";
 import React from "react";
-import { WithTooltip } from "../badge";
 import { KubeObjectAge } from "../kube-object/age";
 import { KubeObjectListLayout } from "../kube-object-list-layout";
 import { KubeObjectStatusIcon } from "../kube-object-status-icon";
 import { SiblingsInTabLayout } from "../layout/siblings-in-tab-layout";
 import { NamespaceSelectBadge } from "../namespaces/namespace-select-badge";
+import { WithTooltip } from "../with-tooltip";
 import limitRangeStoreInjectable from "./store.injectable";
 
 import type { LimitRangeStore } from "./store";
