+++ packages/core/src/renderer/components/config-runtime-classes/runtime-classes.tsx
@@ -10,11 +10,11 @@ import { withInjectables } from "@ogre-tools/injectable-react";
 import autoBindReact from "auto-bind/react";
 import { observer } from "mobx-react";
 import * as React from "react";
-import { WithTooltip } from "../badge";
 import { KubeObjectAge } from "../kube-object/age";
 import { KubeObjectListLayout } from "../kube-object-list-layout";
 import { KubeObjectStatusIcon } from "../kube-object-status-icon";
 import { SiblingsInTabLayout } from "../layout/siblings-in-tab-layout";
+import { WithTooltip } from "../with-tooltip";
 import runtimeClassStoreInjectable from "./store.injectable";
 
 import type { RuntimeClass } from "@freelensapp/kube-object";
