+++ packages/core/src/renderer/components/config-secrets/secrets.tsx
@@ -9,12 +9,13 @@ import "./secrets.scss";
 import { withInjectables } from "@ogre-tools/injectable-react";
 import { observer } from "mobx-react";
 import React from "react";
-import { Badge, WithTooltip } from "../badge";
+import { Badge } from "../badge";
 import { KubeObjectAge } from "../kube-object/age";
 import { KubeObjectListLayout } from "../kube-object-list-layout";
 import { KubeObjectStatusIcon } from "../kube-object-status-icon";
 import { SiblingsInTabLayout } from "../layout/siblings-in-tab-layout";
 import { NamespaceSelectBadge } from "../namespaces/namespace-select-badge";
+import { WithTooltip } from "../with-tooltip";
 import openAddSecretDialogInjectable from "./add-dialog/open.injectable";
 import { AddSecretDialog } from "./add-dialog/view";
 import secretStoreInjectable from "./store.injectable";
