+++ packages/core/src/renderer/components/config-validating-webhook-configurations/validating-webhook-configurations.tsx
@@ -9,10 +9,10 @@ import "./validating-webhook-configurations.scss";
 import { withInjectables } from "@ogre-tools/injectable-react";
 import { observer } from "mobx-react";
 import React from "react";
-import { WithTooltip } from "../badge";
 import { KubeObjectAge } from "../kube-object/age";
 import { KubeObjectListLayout } from "../kube-object-list-layout";
 import { SiblingsInTabLayout } from "../layout/siblings-in-tab-layout";
+import { WithTooltip } from "../with-tooltip";
 import validatingWebhookConfigurationsStoreInjectable from "./validating-webhook-configuration-store.injectable";
 
 import type { ValidatingWebhookConfigurationStore } from "./validating-webhook-configuration-store";
