+++ packages/core/src/renderer/components/helm-charts/helm-charts.tsx
@@ -11,9 +11,9 @@ import { noop } from "lodash";
 import { observer } from "mobx-react";
 import React, { Component } from "react";
 import navigateToHelmChartsInjectable from "../../../common/front-end-routing/routes/cluster/helm/charts/navigate-to-helm-charts.injectable";
-import { WithTooltip } from "../badge";
 import { ItemListLayout } from "../item-object-list/list-layout";
 import { SiblingsInTabLayout } from "../layout/siblings-in-tab-layout";
+import { WithTooltip } from "../with-tooltip";
 import { HelmChartDetails } from "./helm-chart-details";
 import helmChartsInjectable from "./helm-charts/helm-charts.injectable";
 import selectedHelmChartInjectable from "./helm-charts/selected-helm-chart.injectable";
