+++ packages/core/src/renderer/components/helm-releases/releases.tsx
@@ -12,11 +12,11 @@ import { kebabCase } from "lodash/fp";
 import moment from "moment-timezone";
 import React, { Component } from "react";
 import navigateToHelmReleasesInjectable from "../../../common/front-end-routing/routes/cluster/helm/releases/navigate-to-helm-releases.injectable";
-import { WithTooltip } from "../badge";
 import { ItemListLayout } from "../item-object-list";
 import { SiblingsInTabLayout } from "../layout/siblings-in-tab-layout";
 import { NamespaceSelectBadge } from "../namespaces/namespace-select-badge";
 import { NamespaceSelectFilter } from "../namespaces/namespace-select-filter";
+import { WithTooltip } from "../with-tooltip";
 import { ReleaseRollbackDialog } from "./dialog/dialog";
 import helmReleasesRouteParametersInjectable from "./helm-releases-route-parameters.injectable";
 import { HelmReleaseMenu } from "./release-menu";
