+++ packages/core/src/renderer/components/network-endpoint-slices/endpoint-slices.tsx
@@ -9,11 +9,11 @@ import "./endpoint-slices.scss";
 import { withInjectables } from "@ogre-tools/injectable-react";
 import { observer } from "mobx-react";
 import React from "react";
-import { WithTooltip } from "../badge";
 import { KubeObjectAge } from "../kube-object/age";
 import { KubeObjectListLayout } from "../kube-object-list-layout";
 import { SiblingsInTabLayout } from "../layout/siblings-in-tab-layout";
 import { NamespaceSelectBadge } from "../namespaces/namespace-select-badge";
+import { WithTooltip } from "../with-tooltip";
 import endpointSliceStoreInjectable from "./store.injectable";
 
 import type { EndpointSliceStore } from "./store";
