+++ packages/core/src/renderer/components/network-port-forwards/port-forwards.tsx
@@ -12,10 +12,10 @@ import { disposeOnUnmount, observer } from "mobx-react";
 import React from "react";
 import navigateToPortForwardsInjectable from "../../../common/front-end-routing/routes/cluster/network/port-forwards/navigate-to-port-forwards.injectable";
 import portForwardStoreInjectable from "../../port-forward/port-forward-store/port-forward-store.injectable";
-import { WithTooltip } from "../badge";
 import { ItemListLayout } from "../item-object-list/list-layout";
 import { SiblingsInTabLayout } from "../layout/siblings-in-tab-layout";
 import { NamespaceSelectBadge } from "../namespaces/namespace-select-badge";
+import { WithTooltip } from "../with-tooltip";
 import { PortForwardDetails } from "./port-forward-details";
 import { PortForwardMenu } from "./port-forward-menu";
 import portForwardsRouteParametersInjectable from "./port-forwards-route-parameters.injectable";
