+++ packages/core/src/renderer/components/user-management/role-bindings/view.tsx
@@ -9,12 +9,12 @@ import "./view.scss";
 import { withInjectables } from "@ogre-tools/injectable-react";
 import { observer } from "mobx-react";
 import React from "react";
-import { WithTooltip } from "../../badge";
 import { KubeObjectAge } from "../../kube-object/age";
 import { KubeObjectListLayout } from "../../kube-object-list-layout";
 import { KubeObjectStatusIcon } from "../../kube-object-status-icon";
 import { SiblingsInTabLayout } from "../../layout/siblings-in-tab-layout";
 import { NamespaceSelectBadge } from "../../namespaces/namespace-select-badge";
+import { WithTooltip } from "../../with-tooltip";
 import clusterRoleStoreInjectable from "../cluster-roles/store.injectable";
 import roleStoreInjectable from "../roles/store.injectable";
 import serviceAccountStoreInjectable from "../service-accounts/store.injectable";
