+++ packages/core/src/renderer/components/with-tooltip/index.ts
@@ -4,4 +4,4 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 
-export * from "./badge";
+export * from "./with-tooltip";
