+++ packages/core/src/renderer/components/workloads-cronjobs/cronjobs.tsx
@@ -11,13 +11,13 @@ import cronstrue from "cronstrue";
 import { observer } from "mobx-react";
 import moment from "moment-timezone";
 import React from "react";
-import { WithTooltip } from "../badge";
 import eventStoreInjectable from "../events/store.injectable";
 import { KubeObjectAge } from "../kube-object/age";
 import { KubeObjectListLayout } from "../kube-object-list-layout";
 import { KubeObjectStatusIcon } from "../kube-object-status-icon";
 import { SiblingsInTabLayout } from "../layout/siblings-in-tab-layout";
 import { NamespaceSelectBadge } from "../namespaces/namespace-select-badge";
+import { WithTooltip } from "../with-tooltip";
 import cronJobStoreInjectable from "./store.injectable";
 
 import type { EventStore } from "../events/store";
