+++ packages/core/src/renderer/components/workloads-deployments/deployments.tsx
@@ -12,13 +12,13 @@ import kebabCase from "lodash/kebabCase";
 import orderBy from "lodash/orderBy";
 import { observer } from "mobx-react";
 import React from "react";
-import { WithTooltip } from "../badge";
 import eventStoreInjectable from "../events/store.injectable";
 import { KubeObjectAge } from "../kube-object/age";
 import { KubeObjectListLayout } from "../kube-object-list-layout";
 import { KubeObjectStatusIcon } from "../kube-object-status-icon";
 import { SiblingsInTabLayout } from "../layout/siblings-in-tab-layout";
 import { NamespaceSelectBadge } from "../namespaces/namespace-select-badge";
+import { WithTooltip } from "../with-tooltip";
 import deploymentStoreInjectable from "./store.injectable";
 
 import type { Deployment } from "@freelensapp/kube-object";
