+++ packages/core/src/renderer/components/workloads-jobs/jobs.tsx
@@ -10,7 +10,7 @@ import { formatDuration } from "@freelensapp/utilities/dist";
 import { withInjectables } from "@ogre-tools/injectable-react";
 import { observer } from "mobx-react";
 import React from "react";
-import { Badge, WithTooltip } from "../badge";
+import { Badge } from "../badge";
 import eventStoreInjectable from "../events/store.injectable";
 import { KubeObjectAge } from "../kube-object/age";
 import { KubeObjectListLayout } from "../kube-object-list-layout";
@@ -18,6 +18,7 @@ import { KubeObjectStatusIcon } from "../kube-object-status-icon";
 import { SiblingsInTabLayout } from "../layout/siblings-in-tab-layout";
 import { LocaleDate } from "../locale-date";
 import { NamespaceSelectBadge } from "../namespaces/namespace-select-badge";
+import { WithTooltip } from "../with-tooltip";
 import jobStoreInjectable from "./store.injectable";
 
 import type { Job } from "@freelensapp/kube-object";
