+++ packages/core/src/renderer/components/workloads-pods/columns/pods-name-column.injectable.tsx
@@ -8,7 +8,7 @@ import { podListLayoutColumnInjectionToken } from "@freelensapp/list-layout";
 import { getConvertedParts } from "@freelensapp/utilities";
 import { getInjectable } from "@ogre-tools/injectable";
 import React from "react";
-import { WithTooltip } from "../../badge";
+import { WithTooltip } from "../../with-tooltip";
 import { COLUMN_PRIORITY } from "./column-priority";
 
 import type { Pod } from "@freelensapp/kube-object";
